v0.4.2 — Lint-Befund behoben, tools/lint.sh prueft wirklich
Die CI von v0.4.1 war rot: clang-tidy meldete in SlotDisplay.cpp eine verengende
Umwandlung (der Bedingungsoperator rechnet in int, der Cast stand nur um einen Zweig).
Behoben, das Verhalten aendert sich nicht.

Warum es lokal nicht auffiel: tools/lint.sh rief 'pio check --skip-packages'. Das nahm
clang-tidy auch die Include-Pfade des Frameworks; jede Datei brach mit 'Arduino.h file
not found' ab, PlatformIO meldete trotzdem 'No defects found', und der Lauf war in
zweieinhalb Sekunden durch. Eine Pruefung, die nichts prueft und gruen meldet, ist
schlimmer als keine. Jetzt --pattern src --pattern include (Pfade bleiben, nur eigene
Quellen werden analysiert, rund 20 Sekunden) und der Check-Cache wird vorher verworfen.

+++ firmware/src/slots/SlotDisplay.cpp
@@ -172,8 +172,10 @@ void kachelRechteck(const Config& cfg, uint8_t seite, PageLayout l, uint8_t pos,
                                                      inhaltHoeheAufPlatz(cfg, seite, 1),
                                                      cfg.teilung[seite - 1]);
             x = 0;
-            y = (pos == 0) ? 0 : obenH;
-            h = (pos == 0) ? obenH : (int16_t)(HOEHE - obenH);
+            // Der Bedingungsoperator rechnet in int; ohne den Cast um den GANZEN
+            // Ausdruck ist die Zuweisung eine verengende Umwandlung (clang-tidy).
+            y = (int16_t)((pos == 0) ? 0 : obenH);
+            h = (int16_t)((pos == 0) ? obenH : (HOEHE - obenH));
             break;
         }
         default:
